+++ lib/extensions.dart
@@ -1,4 +1,6 @@
-// Copyright ©2022, GM Consult (Pty) Ltd. All rights reserved
+// BSD 3-Clause License
+// Copyright (c) 2022, GM Consult Pty Ltd
+// All rights reserved
 
 /// Exports the extension methods exposed by this package. Also exports
 /// the extensions from the `porter_2_stemmer` package.
@@ -8,3 +10,4 @@ export 'src/term_similarity.dart' show TermSimilarityExtensions;
 export 'src/token.dart' show TokenCollectionExtension;
 export 'package:porter_2_stemmer/porter_2_stemmer.dart'
     show Porter2StemmerExtension;
+export 'src/english_analyzer/english.dart' show EnglishStringExtensions;