+++ lib/src/text_tokenizer.dart
@@ -3,7 +3,6 @@
 // All rights reserved
 
 import '_index.dart';
-import 'package:porter_2_stemmer/porter_2_stemmer.dart';
 
 /// Interface for a text analyser class that extracts tokens from text for use
 /// in full-text search queries and indexes:
@@ -18,14 +17,16 @@ import 'package:porter_2_stemmer/porter_2_stemmer.dart';
 abstract class TextTokenizer {
   //
 
+  /// Returns a static const [TextTokenizer] using the [English] analyzer.
+  static const english = _TextTokenizerImpl(English.analyzer, null);
+
   /// Instantiates a const [TextTokenizerBase] instance.
   /// - [analyzer] is used by the [TextTokenizer] to tokenize source text
   ///   (default is [English.analyzer]); and
   /// - provide a custom [tokenFilter] if you want to manipulate tokens or
   ///   restrict tokenization to tokens that meet specific criteria.
   factory TextTokenizer(
-          {TextAnalyzer analyzer = English.analyzer,
-          TokenFilter? tokenFilter}) =>
+          {required TextAnalyzer analyzer, TokenFilter? tokenFilter}) =>
       _TextTokenizerImpl(analyzer, tokenFilter);
 
   // /// Splits the [source] into paragraphs at line ending marks.
@@ -108,45 +109,10 @@ abstract class TextTokenizerMixin implements TextTokenizer {
     await Future.forEach(terms, (String term) async {
       // remove white-space at start and end of term
       term = term.trim();
-      // only tokenize non-empty strings.
-      if (term.isNotEmpty && !analyzer.stopWords.contains(term)) {
-        // check for exceptions
-        var exception = analyzer.termExceptions[term];
-        if (exception != null) {
-          if (!analyzer.stopWords.contains(exception)) {
-            tokens.add(Token(exception.trim(), position, zone));
-          }
-        } else {
-          // apply the termFilter
-          final splitTerms = await analyzer.termFilter(term);
-          for (var splitTerm in splitTerms) {
-            // check splitTerm is not in exceptions
-            exception = analyzer.termExceptions[splitTerm.trim()]?.trim();
-            // var tokenTerm = splitTerm;
-            if (exception == null || exception.isEmpty) {
-              splitTerm = analyzer
-                  .stemmer(analyzer.lemmatizer(splitTerm.trim()))
-                  .trim();
-              exception = analyzer.termExceptions[splitTerm];
-            }
-            // check splitTerm is not in exceptions
-            if (exception != null && exception.isNotEmpty) {
-              if (!analyzer.stopWords.contains(exception)) {
-                // tokenize the exception
-                tokens.add(Token(exception, position, zone));
-              }
-            } else {
-              if (splitTerm.isNotEmpty) {
-                if (!analyzer.stopWords.contains(exception)) {
-                  // tokenize splitTerm
-                  tokens.add(Token(splitTerm.trim(), position, zone));
-                }
-              }
-            }
-          }
-        }
+      final splitTerms = await analyzer.termFilter(term);
+      for (var splitTerm in splitTerms) {
+        tokens.add(Token(splitTerm.trim(), position, zone));
       }
-      // increment the token position
       position++;
     });
     // apply the tokenFilter if it is not null and return the tokens collection