+++ lib/text_analysis.dart
@@ -1,10 +1,12 @@
-// Copyright ©2022, GM Consult (Pty) Ltd. All rights reserved
+// BSD 3-Clause License
+// Copyright (c) 2022, GM Consult Pty Ltd
+// All rights reserved
 
 /// DART text analyzer that extracts tokens from JSON documents for use in
 /// information retrieval systems.
 library text_analysis;
 
-export 'src/english.dart' show English;
+export 'src/english_analyzer/english.dart' show English;
 export 'src/term_similarity.dart' show TermSimilarity;
 export 'src/text_analyzer.dart' show TextAnalyzer;
 export 'src/text_document.dart'