+++ lib/type_definitions.dart
@@ -1,4 +1,6 @@
-// Copyright ©2022, GM Consult (Pty) Ltd. All rights reserved
+// BSD 3-Clause License
+// Copyright (c) 2022, GM Consult Pty Ltd
+// All rights reserved
 
 /// Exports all the type definitions used in the `text_analysis` library.
 library type_definitions;