+++ test/vocabulary.dart
@@ -1,30 +1,30 @@
-// BSD 3-Clause License
-// Copyright (c) 2022, GM Consult Pty Ltd
+// // BSD 3-Clause License
+// // Copyright (c) 2022, GM Consult Pty Ltd
 
-@Timeout(Duration(seconds: 600))
+// @Timeout(Duration(seconds: 600))
 
-import 'dart:io';
-import 'package:porter_2_stemmer/porter_2_stemmer.dart';
-import 'package:test/test.dart';
-import 'data/terms.dart';
+// import 'dart:io';
+// import 'package:porter_2_stemmer/porter_2_stemmer.dart';
+// import 'package:test/test.dart';
+// import 'data/terms.dart';
 
-void main() {
-  group('vocabulary', () {
-//
+// void main() {
+//   group('vocabulary', () {
+// //
 
-    test('stem vocab', () async {
-      final out = File('vocabulary.txt').openWrite();
-      out.write('const vocabulary = [\n');
-      // final analyzer = TextTokenizer();
-      for (var e in terms) {
-        // final splitTerms = await analyzer.configuration.termFilter(term);
-        e = e.stemPorter2();
-        final entry = '"$e",\n';
-        out.write(entry);
-      }
+//     test('stem vocab', () async {
+//       final out = File('vocabulary.txt').openWrite();
+//       out.write('const vocabulary = [\n');
+//       // final analyzer = TextTokenizer();
+//       for (var e in terms) {
+//         // final splitTerms = await analyzer.configuration.termFilter(term);
+//         e = e.stemPorter2();
+//         final entry = '"$e",\n';
+//         out.write(entry);
+//       }
 
-      out.write('];\n');
-      await out.close();
-    });
-  });
-}
+//       out.write('];\n');
+//       await out.close();
+//     });
+//   });
+// }